@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,8 @@ import PackageDescription
 let package = Package(
     name: "T2ScholaCoreSwift",
     platforms: [
-            .macOS(.v10_12),
+            .macOS(.v12),
+            .iOS(.v15)
         ],
     products: [
         .library(
