@@ -14,92 +14,36 @@ public struct T2Schola {
         self.apiClient = apiClient
     }
     
-    public func getToken(completionHandler: @escaping (Result<String, Error>) -> Void) {
-        apiClient.send(request: LoginRequest()) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response.wsToken))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func getToken() async throws -> String {
+        try await apiClient.send(request: LoginRequest()).wsToken
     }
     
-    public func getSiteInfo(wsToken: String, completionHandler: @escaping (Result<SiteInfoResponse, Error>) -> Void) {
-        apiClient.send(request: SiteInfoRequest(wsToken: wsToken)) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func getSiteInfo(wsToken: String) async throws -> SiteInfoResponse {
+        try await apiClient.send(request: SiteInfoRequest(wsToken: wsToken))
     }
     
-    public func getUserCourses(userId: Int, wsToken: String, completionHandler: @escaping (Result<UserEnrolCoursesResponse, Error>) -> Void) {
-        apiClient.send(request: UserEnrolCoursesRequest(userId: userId, wsToken: wsToken)) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func getUserCourses(userId: Int, wsToken: String) async throws -> UserEnrolCoursesResponse {
+        try await apiClient.send(request: UserEnrolCoursesRequest(userId: userId, wsToken: wsToken))
     }
     
-    public func getCourseCategories(wsToken: String, completionHandler: @escaping (Result<CourseCategoriesResponse, Error>) -> Void) {
-        apiClient.send(request: CourseCategoriesRequest(wsToken: wsToken)) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func getCourseCategories(wsToken: String) async throws -> CourseCategoriesResponse {
+        try await apiClient.send(request: CourseCategoriesRequest(wsToken: wsToken))
     }
     
-    public func getCourseContents(courseId: Int, wsToken: String, completionHandler: @escaping (Result<CourseContentsResponse, Error>) -> Void) {
-        apiClient.send(request: CourseContentsRequest(courseid: courseId, wsToken: wsToken)) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func getCourseContents(courseId: Int, wsToken: String) async throws -> CourseContentsResponse {
+        try await apiClient.send(request: CourseContentsRequest(courseid: courseId, wsToken: wsToken))
     }
     
-    public func getAssignments(wsToken: String, completionHandler: @escaping (Result<AssignmentsResponse, Error>) -> Void) {
-        apiClient.send(request: AssignmentsRequest(wsToken: wsToken)) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func getAssignments(wsToken: String) async throws -> AssignmentsResponse {
+        try await apiClient.send(request: AssignmentsRequest(wsToken: wsToken))
     }
     
-    public func getAssignmentSubmissionStatus(assignmentId: Int, userId: Int, wsToken: String, completionHandler: @escaping (Result<AssignmentSubmissionStatusResponse, Error>) -> Void) {
-        apiClient.send(request: AssignmentSubmissionStatusRequest(assignmentId: assignmentId, userId: userId, wsToken: wsToken)) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func getAssignmentSubmissionStatus(assignmentId: Int, userId: Int, wsToken: String) async throws -> AssignmentSubmissionStatusResponse {
+        try await apiClient.send(request: AssignmentSubmissionStatusRequest(assignmentId: assignmentId, userId: userId, wsToken: wsToken))
     }
     
-    public func updateActivityCompletionStatusManually(moduleId: Int, completed: Bool, wsToken: String, completionHandler: @escaping (Result<UpdateActivityCompletionStatusManuallyResponse, Error>) -> Void) {
-        apiClient.send(request: UpdateActivityCompletionStatusManuallyRequest(moduleId: moduleId, completed: completed, wsToken: wsToken)) { result in
-            switch result {
-            case let .success(response):
-                completionHandler(.success(response))
-            case let .failure(error):
-                completionHandler(.failure(error))
-            }
-        }
+    public func updateActivityCompletionStatusManually(moduleId: Int, completed: Bool, wsToken: String) async throws -> UpdateActivityCompletionStatusManuallyResponse {
+        try await apiClient.send(request: UpdateActivityCompletionStatusManuallyRequest(moduleId: moduleId, completed: completed, wsToken: wsToken))
     }
     
     public static func changeToMock() {
